+++ src/cpp/include/tbsla/cpp/MatrixCOO.hpp
@@ -10,16 +10,16 @@ namespace tbsla { namespace cpp {
 class MatrixCOO : public virtual Matrix {
   public:
     friend std::ostream & operator<<( std::ostream &os, const MatrixCOO &m);
-    MatrixCOO(int n_row, int n_col, double* values, int* row, int* col);
-    MatrixCOO(int n_row, int n_col, long int n_values);
-    MatrixCOO(int n_row, int n_col);
+    MatrixCOO(long long int n_row, long long int n_col, double* values, long long int* row, long long int* col);
+    MatrixCOO(long long int n_row, long long int n_col, long long int n_values);
+    MatrixCOO(long long int n_row, long long int n_col);
     MatrixCOO() : values(0), row(0), col(0) {};
     ~MatrixCOO();
     const double* get_values() const { return values; }
-    const int* get_row() const { return row; }
-    const int* get_col() const { return col; }
-    double* spmv(const double* v, int vect_incr = 0) const;
-    inline void Ax(double* r, const double* v, int vect_incr = 0) const;
+    const long long int* get_row() const { return row; }
+    const long long int* get_col() const { return col; }
+    double* spmv(const double* v, long long int vect_incr = 0) const;
+    inline void Ax(double* r, const double* v, long long int vect_incr = 0) const;
     using tbsla::cpp::Matrix::a_axpx_;
     using tbsla::cpp::Matrix::AAxpAx;
     using tbsla::cpp::Matrix::AAxpAxpx;
@@ -30,11 +30,11 @@ class MatrixCOO : public virtual Matrix {
     void NUMAinit();
 
     void readMM(std::string name);
-    void fill_cdiag(int n_row, int n_col, int cdiag, int pr = 0, int pc = 0, int NR = 1, int NC = 1);
-    void fill_cqmat(int n_row, int n_col, int c, double q, unsigned int seed_mult = 1, int pr = 0, int pc = 0, int NR = 1, int NC = 1);
-    void fill_random(int n_row, int n_col, double nnz_ratio, unsigned int seed_mult = 1, int pr = 0, int pc = 0, int NR = 1, int NC = 1);
-    void fill_cdistrib(int n_row, int n_col, int nnz, int pr = 0, int pc = 0, int NR = 1, int NC = 1);
-    void fill_brain(int n_row, int n_col, int* neuron_type, std::vector<std::vector<double> > proba_conn, std::vector<std::unordered_map<int,std::vector<int> > > brain_struct, unsigned int seed_mult, int pr = 0, int pc = 0, int NR = 1, int NC = 1);
+    void fill_cdiag(long long int n_row, long long int n_col, long long int cdiag, long long int pr = 0, long long int pc = 0, long long int NR = 1, long long int NC = 1);
+    void fill_cqmat(long long int n_row, long long int n_col, long long int c, double q, unsigned long long int seed_mult = 1, long long int pr = 0, long long int pc = 0, long long int NR = 1, long long int NC = 1);
+    void fill_random(long long int n_row, long long int n_col, double nnz_ratio, unsigned long long int seed_mult = 1, long long int pr = 0, long long int pc = 0, long long int NR = 1, long long int NC = 1);
+    void fill_cdistrib(long long int n_row, long long int n_col, long long int nnz, long long int pr = 0, long long int pc = 0, long long int NR = 1, long long int NC = 1);
+    void fill_brain(long long int n_row, long long int n_col, int* neuron_type, std::vector<std::vector<double> > proba_conn, std::vector<std::unordered_map<int,std::vector<int> > > brain_struct, unsigned long long int seed_mult, long long int pr = 0, long long int pc = 0, long long int NR = 1, long long int NC = 1);
 	
 	void get_row_sums(double* buffer);
 	void normalize_rows(double* buffer);
@@ -44,11 +44,11 @@ class MatrixCOO : public virtual Matrix {
 
   protected:
     double* values;
-    int* row;
-    int* col;
+    long long int* row;
+    long long int* col;
 
   private:
-    void init(int n_row, int n_col, long int n_values);
+    void init(long long int n_row, long long int n_col, long long int n_values);
 };
 
 }}
