+++ src/cpp/interface/c/include/tbsla/cpp/MatrixDENSE.h
@@ -13,9 +13,9 @@ typedef struct C_CPP_MatrixDENSE C_CPP_MatrixDENSE_t;
 C_CPP_MatrixDENSE_t *C_CPP_MatrixDENSE_create();
 void C_CPP_MatrixDENSE_destroy(C_CPP_MatrixDENSE_t *m);
 
-void C_CPP_MatrixDENSE_fill_cdiag(C_CPP_MatrixDENSE_t *m, int n_row, int n_col, int cdiag, int pr, int pc, int NR, int NC);
-void C_CPP_MatrixDENSE_fill_cqmat(C_CPP_MatrixDENSE_t *m, int n_row, int n_col, int c, double q, unsigned int seed_mult, int pr, int pc, int NR, int NC);
-bool C_CPP_MatrixDENSE_read(C_CPP_MatrixDENSE_t *m, char *filename, int seek);
+void C_CPP_MatrixDENSE_fill_cdiag(C_CPP_MatrixDENSE_t *m, long long int n_row, long long int n_col, long long int cdiag, long long int pr, long long int pc, long long int NR, long long int NC);
+void C_CPP_MatrixDENSE_fill_cqmat(C_CPP_MatrixDENSE_t *m, long long int n_row, long long int n_col, long long int c, double q, unsigned long long int seed_mult, long long int pr, long long int pc, long long int NR, long long int NC);
+bool C_CPP_MatrixDENSE_read(C_CPP_MatrixDENSE_t *m, char *filename, long long int seek);
 bool C_CPP_MatrixDENSE_write(C_CPP_MatrixDENSE_t *m, char *filename);
 void C_CPP_MatrixDENSE_print(C_CPP_MatrixDENSE_t *m);
 C_CPP_Vector_t *C_CPP_MatrixDENSE_spmv(C_CPP_MatrixDENSE_t *m, C_CPP_Vector_t *v);
